- fix: rename widget names
