+++ frontend/src/widgets/PointCloud2DViewerV2Widget/AngleIndicator.tsx
