+++ frontend/src/widgets/PointCloud2DViewerV2Widget/ColorModeSelector.tsx
