+++ frontend/src/widgets/PointCloud2DViewerV2Widget/PointCloud2DViewerV2Widget.tsx
@@ -11,9 +11,12 @@ import type { WidgetProps } from "@/widgets/index.ts";
 
 import { WidgetFrame } from "@/components/Dashboard/WidgetFrame.tsx";
 import { useMosaicStore } from "@/hooks/useMosaicStore.ts";
-import AngleIndicator from "@/widgets/PPC2DViewerWidget/AngleIndicator.tsx";
 
-export default function PPC2DViewerWidget({ widgetConfig }: WidgetProps) {
+import AngleIndicator from "./AngleIndicator.tsx";
+import { calculateColor, calculatePointSize, type ColorMode } from "./colorMapping.ts";
+import ColorModeSelector from "./ColorModeSelector.tsx";
+
+export default function PointCloud2DViewerV2Widget({ widgetConfig }: WidgetProps) {
   const { getOrCreateStore, releaseStore } = useMosaicStore();
   const canvasRef = useRef<HTMLCanvasElement>(null);
   const [lastUpdate, setLastUpdate] = useState<Date | null>(null);
@@ -23,6 +26,7 @@ export default function PPC2DViewerWidget({ widgetConfig }: WidgetProps) {
   const canvasSizeRef = useRef({ width: 0, height: 0 });
   const lastPPCMetaRef = useRef<PointCloudMeta | null>(null);
   const lastPPCPointsRef = useRef<PointCloudPoint[] | null>(null);
+  const [colorMode, setColorMode] = useState<ColorMode>("height"); // Default to height-based coloring
 
   const clearCanvas = (canvas: HTMLCanvasElement, ctx: CanvasRenderingContext2D) => {
     ctx.fillStyle = "black";
@@ -71,11 +75,13 @@ export default function PPC2DViewerWidget({ widgetConfig }: WidgetProps) {
         setPointCount(meta.height * meta.width);
 
         if (lastPPCPointsRef.current === null) {
+          // First chunk of new frame - draw all points
           lastPPCPointsRef.current = points;
-          drawPoint(canvas, ctx, points, meta);
+          drawPoint(canvas, ctx, points, meta, false);
         } else {
+          // Subsequent chunks - only draw new points
+          drawPoint(canvas, ctx, points, meta, true);
           lastPPCPointsRef.current = [...lastPPCPointsRef.current, ...points];
-          drawPoint(canvas, ctx, lastPPCPointsRef.current, meta);
         }
       } catch (error) {
         console.error("Error processing PointCloud2:", error);
@@ -98,6 +104,7 @@ export default function PPC2DViewerWidget({ widgetConfig }: WidgetProps) {
     ctx: CanvasRenderingContext2D,
     points: PointCloudPoint[],
     meta: PointCloudMeta,
+    incremental = false, // If true, only render new points without clearing
   ) => {
     // Get actual canvas display size and set internal resolution
     const rect = canvas.getBoundingClientRect();
@@ -112,6 +119,8 @@ export default function PPC2DViewerWidget({ widgetConfig }: WidgetProps) {
       canvas.width = canvasWidth;
       canvas.height = canvasHeight;
       canvasSizeRef.current = { width: canvasWidth, height: canvasHeight };
+      // Clear canvas when resizing
+      clearCanvas(canvas, ctx);
     }
 
     // Check minimum size
@@ -119,24 +128,25 @@ export default function PPC2DViewerWidget({ widgetConfig }: WidgetProps) {
       return;
     }
 
-    // Initialize depth map, height map, and intensity map for 2D projection
-    const n = canvasWidth * canvasHeight;
-    const depthMap = Array.from<number>({ length: n }).fill(Number.POSITIVE_INFINITY);
-    const heightMap = Array.from<number>({ length: n }).fill(0);
+    // Calculate max distance for point size calculation
+    const maxDistance = Math.sqrt(meta.max_x * meta.max_x + meta.max_y * meta.max_y);
 
-    let minHeight = meta.min_z;
-    let maxHeight = meta.max_z;
-
-    // Set default values
-    const heightRange = maxHeight - minHeight;
-    if (heightRange === 0) {
-      minHeight = -1;
-      maxHeight = 1;
+    // Depth buffer to handle occlusion
+    // Store points with their screen coordinates and depth for rendering
+    interface PointData {
+      x: number;
+      y: number;
+      distance: number;
+      point: PointCloudPoint;
     }
 
-    // Actual projection and mapping
+    const pointsToRender: PointData[] = [];
+
+    // Project all points to screen space
     for (let i = 0; i < points.length; i++) {
-      const { x, y, z } = points[i];
+      const point = points[i];
+      const { x, y, z } = point;
+
       if (x === null || y === null || z === null) {
         continue;
       }
@@ -144,42 +154,59 @@ export default function PPC2DViewerWidget({ widgetConfig }: WidgetProps) {
       const distance = Math.sqrt(x * x + y * y);
       if (distance === 0) continue;
 
+      // Cylindrical projection
       const azimuth = Math.atan2(y, x);
-      const u = Math.floor(((azimuth + Math.PI) / (2 * Math.PI)) * canvasWidth);
+      const screenX = ((azimuth + Math.PI) / (2 * Math.PI)) * canvasWidth;
 
       // Map height to canvas Y coordinate (top is higher elevation)
-      const v = Math.floor(((maxHeight - z) / (maxHeight - minHeight)) * (canvasHeight - 1));
+      const heightRange = meta.max_z - meta.min_z || 1;
+      const screenY = ((meta.max_z - z) / heightRange) * canvasHeight;
 
-      // Range check
-      if (u >= 0 && u < canvasWidth && v >= 0 && v < canvasHeight) {
-        const pixelIndex = v * canvasWidth + u;
-        // Use closest distance when multiple points map to same pixel
-        if (distance < depthMap[pixelIndex]) {
-          depthMap[pixelIndex] = distance;
-          heightMap[pixelIndex] = z;
-        }
-      }
+      pointsToRender.push({
+        x: screenX,
+        y: screenY,
+        distance,
+        point,
+      });
     }
 
-    // Create image data
-    const imageData = ctx.createImageData(canvas.width, canvas.height);
-
-    // Convert depth map, height map, and intensity map to image
-    for (let pixelIndex = 0; pixelIndex < depthMap.length; pixelIndex++) {
-      if (depthMap[pixelIndex] !== Number.POSITIVE_INFINITY) {
-        // Set color values in image data
-        const imageIndex = pixelIndex * 4;
-        imageData.data[imageIndex] = 255; // R
-        imageData.data[imageIndex + 1] = 255; // G
-        imageData.data[imageIndex + 2] = 255; // B
-        imageData.data[imageIndex + 3] = 255; // A
-      }
+    // Sort by distance (far to near) for proper rendering order
+    // Only sort when not doing incremental rendering for better performance
+    if (!incremental) {
+      pointsToRender.sort((a, b) => b.distance - a.distance);
     }
 
-    // Draw image data to canvas
-    ctx.putImageData(imageData, 0, 0);
+    // Render each point as a circle
+    for (const { x, y, distance, point } of pointsToRender) {
+      // Calculate color based on current mode
+      const [r, g, b, a] = calculateColor(point, meta, colorMode);
+
+      // Calculate point size based on distance
+      const size = calculatePointSize(distance, maxDistance);
+
+      // Set fill style with calculated color
+      ctx.fillStyle = `rgba(${r}, ${g}, ${b}, ${a / 255})`;
+
+      // Draw point as a filled circle
+      ctx.beginPath();
+      ctx.arc(x, y, size / 2, 0, Math.PI * 2);
+      ctx.fill();
+    }
+
     setPointCount(points.length);
     setLastUpdate(new Date());
+  };
+
+  const handleColorModeChange = (mode: ColorMode) => {
+    setColorMode(mode);
+    // Redraw with new color mode
+    if (canvasRef.current && lastPPCPointsRef.current && lastPPCMetaRef.current) {
+      const ctx = canvasRef.current.getContext("2d");
+      if (ctx) {
+        clearCanvas(canvasRef.current, ctx);
+        drawPoint(canvasRef.current, ctx, lastPPCPointsRef.current, lastPPCMetaRef.current, false);
+      }
+    }
   };
 
   const footerInfo = [
@@ -190,6 +217,10 @@ export default function PPC2DViewerWidget({ widgetConfig }: WidgetProps) {
     {
       label: "Last Update",
       value: lastUpdate ? lastUpdate.toLocaleTimeString() : "N/A",
+    },
+    {
+      label: "Mode",
+      value: colorMode.charAt(0).toUpperCase() + colorMode.slice(1),
     },
   ];
 
@@ -210,18 +241,21 @@ export default function PPC2DViewerWidget({ widgetConfig }: WidgetProps) {
           <Box fontSize="sm">{error}</Box>
         </Flex>
       ) : (
-        <>
-          <AngleIndicator height="20px" fontSize="10px" fontColor="gray.300" />
-          <canvas
-            ref={canvasRef}
-            style={{
-              width: "100%",
-              height: "calc(100% - 20px)",
-              objectFit: "contain",
-              borderRadius: "6px",
-            }}
-          />
-        </>
+        <Flex direction="column" h="100%" position="relative">
+          <ColorModeSelector colorMode={colorMode} onChange={handleColorModeChange} />
+          <AngleIndicator />
+          <Box flex="1" position="relative">
+            <canvas
+              ref={canvasRef}
+              style={{
+                width: "100%",
+                height: "100%",
+                objectFit: "contain",
+                borderRadius: "6px",
+              }}
+            />
+          </Box>
+        </Flex>
       )}
     </WidgetFrame>
   );
