+++ frontend/src/widgets/PointCloud2DViewerV2Widget/colorMapping.ts
