+++ frontend/src/widgets/PointCloud2DViewerWidget/AngleIndicator.tsx
