+++ frontend/src/widgets/PointCloud3DViewerWidget/ColorModeSelector.tsx
