+++ frontend/src/widgets/PointCloud3DViewerWidget/ThreeScene.ts
