+++ frontend/src/widgets/PointCloud3DViewerWidget/ViewControls.tsx
