+++ frontend/src/widgets/PointCloud3DViewerWidget/colorMapping.ts
